Module_2/task_manager/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
psycopg2-binary
PyJWT
websockets